requirements-api.txt
fastapi==0.111.1
huggingface-hub==0.23.5
loguru==0.7.3
numpy==2.1.3
pillow==10.4.0
pydantic==2.12.2
python-multipart==0.0.9
uvicorn[standard]==0.30.6
